MGSC-0001 Added Repository and API call
Created CardApi interface to make an API call. Created
CardRepository to make it possible to add cards from the API call
to the database.

The app still has no user functionality, but in the future this
will allow for data persistence from the database and for actual
data to be collected from the API call.