MGSC-0008 Added functionality for selecting cards
Added CardDetailActivity to launch when a card is selected from the
recyclerview using intents to transfer card data and position.
The new activity will return the position value to set the
recyclerview to the same spot.

Users can now select cards to learn more about them. Pressing back
returns the user to the previous item they were on in the card
list.